backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
black==26.3.0
certifi==2026.2.25
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
diff_cover==10.2.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-magic==0.4.27
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
ruamel.yaml==0.19.1
soupsieve==2.8.3
sqlfluff==4.0.4
starlette==0.52.1
tblib==3.2.2
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
